+++ src/hono/prompt/index.ts
@@ -4,5 +4,11 @@ import { add } from './add';
 import { cancel } from './cancel';
 import { get } from './get';
 import { list } from './list';
+import { purge } from './purge';
 
-export const prompt = new Hono().route('/add', add).route('/get', get).route('/list', list).route('/cancel', cancel);
+export const prompt = new Hono()
+  .route('/add', add)
+  .route('/get', get)
+  .route('/list', list)
+  .route('/cancel', cancel)
+  .route('/purge', purge);