+++ src/lib/config.ts
@@ -8,7 +8,8 @@ export const config = {
     level: envVar.get('LOG_LEVEL').default('info').asString()
   },
   http: {
-    port: envVar.get('PORT').default(3000).asInt()
+    port: envVar.get('PORT').default(3000).asInt(),
+    apiKey: envVar.get('API_KEY').default('').asString()
   },
   database: {
     filename: envVar.get('DATABASE_FILENAME').default('./database.sqlite').asString()
@@ -17,6 +18,7 @@ export const config = {
     url: envVar.get('OPENAI_URL').default('http://localhost:8080/v1').asString(),
     model: envVar.get('OPENAI_MODEL').default('').asString(),
     key: envVar.get('OPENAI_KEY').default('none').asString(),
-    timeout: envVar.get('OPENAI_TIMEOUT').default(10_000).asInt()
+    timeout: envVar.get('OPENAI_TIMEOUT').default(10_000).asInt(),
+    maxRetryCount: envVar.get('OPENAI_MAX_RETRY_COUNT').default(10).asInt()
   }
 };